+++ src/chat/ofxICChatSession.h
@@ -27,7 +27,9 @@ public:
 	ChatResult send(
 		const std::string & message,
 		ChatChunkCallback onChunk = nullptr,
-		std::function<bool()> shouldCancel = nullptr);
+		RequestControl control = {});
+	ChatResult send(const std::string & message, ChatChunkCallback onChunk,
+		std::function<bool()> shouldCancel);
 
 private:
 	friend class ToolLoop;
@@ -36,7 +38,7 @@ private:
 		std::vector<ChatMessage> newMessages,
 		const std::vector<ToolDefinition> & tools,
 		ChatChunkCallback onChunk = nullptr,
-		std::function<bool()> shouldCancel = nullptr);
+		RequestControl control = {});
 
 	std::reference_wrapper<Endpoint> endpoint;
 	std::string systemPrompt;