+++ src/ofxIC.h
@@ -8,6 +8,7 @@
 #include "chat/ofxICChatTypes.h"
 #include "documents/ofxICDocumentIndex.h"
 #include "endpoint/ofxICEndpoint.h"
+#include "endpoint/ofxICRequestTypes.h"
 #include "media/ofxICMediaClient.h"
 #include "media/ofxICMediaTypes.h"
 #include "music/ofxICAceStepMusicClient.h"